requirements.txt
beautifulsoup4==4.12.3
PyJWT==2.9.0
requests==2.31.0